Some updates
Firstly I converted whole program into functional code for better understanding. The bug in Hangman is that if we enter a full word it's getting accepted so what if the user just input whole 26 letters("abcdefghijklmnopqrstuvwxyz") the game over and the player wins and another bug is that it counts miss for the same character input this means if user input 's' which is not there but if again user input 's' it's counting as a wrong input and user will lose one more chance to guess. Bugs in Stone Paper Scissor is that what if the user input "PaPeR" or "YeS" so I had converted the input into a small letter string using lower(). Dice Roll having the same bug as Stone paper game for "YeS". Guess the number having the same bugs as Hangman for string "0123456789" and repeating the wrong digit.